+++ flow/stubs/CSSModuleStub.js
@@ -1,4 +1,0 @@
-// @flow
-type CSSModule = { [key: string]: string };
-const emptyCSSModule: CSSModule = {};
-export default emptyCSSModule;