+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Internal: Detect invalid composes in CSS modules (#676)
 - Internal: Flowtype CSS files in CI (#675)
+- Internal: Add code of conduct & powered by netlify link (#679)
 
 ### Patch